Pod Index has changed! Instead of the arena having an index of its own, every single physical object has its own pod Index. this way, objects like arrows can move smoothly between pods.
I also gave arrows direction and put a placeholder sprite in so that their direction can be easily seen.

The player also has its own pod Index that's used to decide which pod gets rendered. I placed a new DrawImageRot function in Picasso to allow for the rotated arrow sprites.

The arrows also required me to add a "sticky" physical attribute to the physics function so that objects that hit walls and stick to them can be implemented. Right now sticky objects burrow into walls a bit too deep and don't stick well to the underside of walls.

Finally, I changed shooting mechanics from fire on click to a held charged shot, which feels more archery-ish. There is a placeholder charge-indicator bar on the bottom of the screen when LMB is held.